Refactor Plan Limiter service to account for custom plans

+++ src/main/java/org/chainoptim/core/subscription/model/CustomSubscriptionPlan.java
@@ -16,5 +16,5 @@ public class CustomSubscriptionPlan {
     private SubscriptionPlanTier planTier;
     private Boolean isMonthly;
     private Float totalDollarsMonthly;
-    private Map<Feature, Float> additionalFeatures;
+    private Map<Feature, Long> additionalFeatures;
 }
